+++ src/core/scheduler/scheduler.service.ts
@@ -5,5 +5,6 @@ import * as schedule from 'node-schedule';
 export class SchedulerService {
   public readonly scheduleJob = schedule.scheduleJob;
   public readonly scheduledJobs = schedule.scheduledJobs;
+  public readonly cancelJob = schedule.cancelJob;
   public readonly RecurrenceRule = schedule.RecurrenceRule;
 }